requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
alembic
psycopg2-binary
python-dotenv
langchain
